🐛 Fix Guest Portal C# data type conversions for compilation
Fixed all CS1503 compilation errors by updating variable types and
Convert methods to match updated database schema.

**Data type fixes (int → long/byte):**

Admin/RoomQRGenerator.aspx.cs:
- Line 121: int accommodationId → byte accommodationId
- Line 124: Convert.ToInt32 → Convert.ToByte

Guest/Concierge.aspx.cs:
- Fields: int _reservationId → long, int _accommodationId → byte
- Line 30: Convert.ToInt32 → Convert.ToInt64
- Line 32: Convert.ToInt32 → Convert.ToByte

Guest/Housekeeping.aspx.cs:
- Fields: int _reservationId → long, int _accommodationId → byte
- Line 53: Convert.ToInt32 → Convert.ToInt64
- Line 55: Convert.ToInt32 → Convert.ToByte

Guest/Portal.aspx.cs:
- Line 97: int reservationId → long, Convert.ToInt32 → Convert.ToInt64
- Line 98: int accommodationId → byte, Convert.ToInt32 → Convert.ToByte
- Line 99-100: CheckIn/CheckOut → CheckInDate/CheckOutDate

Guest/RoomService.aspx.cs:
- Fields: int _reservationId → long, int _accommodationId → byte
- Line 61: Convert.ToInt32 → Convert.ToInt64
- Line 63: Convert.ToInt32 → Convert.ToByte

All changes ensure proper type matching with:
- Reservation_ID: BIGINT → long
- Accommodation_ID: TINYINT → byte

+++ Take Time BangPhra/Admin/RoomQRGenerator.aspx.cs
@@ -118,7 +118,7 @@ namespace Take_Time_BangPhra.Admin
             {
                 try
                 {
-                    int accommodationId = Convert.ToInt32(e.CommandArgument);
+                    byte accommodationId = Convert.ToByte(e.CommandArgument);
                     short adminId = Session["UserID"] != null ? Convert.ToInt16(Session["UserID"]) : (short)1;
 
                     DataTable result = _guestPortalService.GenerateRoomQRCode(accommodationId, adminId);
